Refactor controllers and services to remove unnecessary autowiring and improve package structure
