+++ centralhub/src/main/java/com/capstonebau2025/centralhub/controller/DeviceController.java
@@ -1,33 +1,14 @@
 package com.capstonebau2025.centralhub.controller;
 
-import com.capstonebau2025.centralhub.entity.Device;
-import com.capstonebau2025.centralhub.service.DeviceService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
+import com.capstonebau2025.centralhub.service.device.DeviceService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/devices")
+@RequiredArgsConstructor
 public class DeviceController {
 
     private final DeviceService deviceService;
-
-    @Autowired
-    public DeviceController(DeviceService deviceService) {
-        this.deviceService = deviceService;
-    }
-
-    // Get Device by ID
-    @GetMapping("/{id}")
-    public ResponseEntity<Device> getDeviceById(@PathVariable long id) {
-        return ResponseEntity.of(deviceService.getById(id));
-    }
-
-    // Delete Device
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteDevice(@PathVariable Long id) {
-        deviceService.deleteById(id);
-        return ResponseEntity.noContent().build();
-    }
 }
