+++ centralhub/src/main/java/com/capstonebau2025/centralhub/controller/EventController.java
@@ -1,26 +1,15 @@
 package com.capstonebau2025.centralhub.controller;
 
-import com.capstonebau2025.centralhub.entity.Event;
-import com.capstonebau2025.centralhub.service.crud.EventService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
+import com.capstonebau2025.centralhub.service.device.EventService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/events")
+@RequiredArgsConstructor
 public class EventController {
 
-    private EventService eventService;
+    private final EventService eventService;
 
-    @Autowired
-    public EventController(EventService eventService) {
-        this.eventService = eventService;
-    }
-
-    // Get Event by ID
-    @GetMapping("/{id}")
-    public ResponseEntity<Event> getEventById(@PathVariable long id) {
-        return ResponseEntity.of(eventService.getById(id));
-    }
 }
