+++ centralhub/src/main/java/com/capstonebau2025/centralhub/service/AreaService.java
@@ -1,0 +1,9 @@
+package com.capstonebau2025.centralhub.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AreaService {
+
+}
+
