+++ centralhub/src/main/java/com/capstonebau2025/centralhub/service/automation/AutomationExecService.java
@@ -1,0 +1,11 @@
+package com.capstonebau2025.centralhub.service.automation;
+
+public class AutomationExecService {
+
+    /*
+    * method for subscribing for events and state changes
+    *
+    * method for executing actions if if event, state change, or time is triggered,
+    * it should be implemented every second or minute
+    * */
+}
