+++ centralhub/src/main/java/com/capstonebau2025/centralhub/service/automation/AutomationService.java
@@ -1,0 +1,21 @@
+package com.capstonebau2025.centralhub.service.automation;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AutomationService {
+
+    /*
+    * method for creating an automation including rule, trigger, and their actions,
+    * and call automationExecService to subscribe for events and state changes
+    *
+    * method for updating automation rule name
+    *
+    * method for updating automation trigger
+    *
+    * method for removing an automation action for an automation
+    *
+    * method for adding an automation action for an automation
+    *
+    * */
+}
