--- centralhub/src/main/java/com/capstonebau2025/centralhub/service/crud/GenericService.java
@@ -1,24 +1,0 @@
-package com.capstonebau2025.centralhub.service.crud;
-// GenericService.java
-
-import java.util.Optional;
-
-public interface GenericService <T, ID> {
-    T create(T entity);
-    T update(T entity);
-    T save(T entity);  // Save will be a generic method that can be used for both create and update
-    Optional<T> getById(ID id);
-    Iterable<T> getAll();
-    void deleteById(ID id);
-    void delete(T entity);
-}
-
-
-
-
-
-
-
-
-
-
