+++ centralhub/src/main/java/com/capstonebau2025/centralhub/service/device/CommandService.java
@@ -1,0 +1,9 @@
+package com.capstonebau2025.centralhub.service.device;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CommandService {
+
+}
+
