+++ centralhub/src/main/java/com/capstonebau2025/centralhub/service/device/EventService.java
@@ -1,0 +1,8 @@
+package com.capstonebau2025.centralhub.service.device;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class EventService {
+
+}
