+++ centralhub/src/main/java/com/capstonebau2025/centralhub/service/device/StateSyncService.java
@@ -1,0 +1,4 @@
+package com.capstonebau2025.centralhub.service.device;
+
+public class StateSyncService {
+}
